requirements.txt
pandas
faiss-cpu
sentence-transformers
python-dotenv
langchain
langgraph
exa-py
langfuse 